@@ -1,12 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { Work } from '../models/work';
+import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout'
+import { Component, type OnInit } from '@angular/core'
+import { type Work } from '../../models/work'
 
 @Component({
   selector: 'app-works',
   templateUrl: './works.component.html',
-  styleUrls: ['./works.component.scss']
+  styleUrl: './works.component.scss'
 })
+
 export class WorksComponent implements OnInit {
+  numberOfColumns: number = 2
   works: Work[] = [
     {
       title: 'Amazon SES email template management page',
@@ -20,13 +23,21 @@ export class WorksComponent implements OnInit {
       screenshot: './assets/myPersonalWebsiteScreenshot.png',
       link: '',
       sourceCodeLink: 'https://github.com/joviefung/jovie-personal-website',
-      framework: 'Angular 8, Typescript'
+      framework: 'Angular 17, Typescript'
     }
   ]
 
-  constructor() { }
+  constructor (private readonly breakpointObserver: BreakpointObserver) {}
 
-  ngOnInit() {
+  ngOnInit () {
+    this.breakpointObserver.observe([Breakpoints.HandsetPortrait])
+      .subscribe(result => {
+        if (result.matches) {
+          this.numberOfColumns = 1
+        } else {
+          this.numberOfColumns = 2
+        }
+      })
   }
 
 }
