@@ -2,6 +2,6 @@ import { TimelineEvent } from './timeline-event';
 
 describe('TimelineEvent', () => {
   it('should create an instance', () => {
-    expect(new TimelineEvent()).toBeTruthy();
+    expect(new TimelineEvent({})).toBeTruthy();
   });
 });
